+++ cogs/general_cog.py
@@ -4,7 +4,7 @@ from discord.ext import commands
 from bot_globals import logger
 from embeds.general_embeds import COMMAND_CATEGORIES, help_embed
 from middleware import defer_interaction, track_analytics
-from utils.views_utils import CommandTypeSelectView
+from views.commands_help_view import CommandTypeSelectView
 
 
 class General(commands.Cog):
